crud de arquitectura funcional con manejo de archivos repetidos

+++ solicitud_asignacion.php
@@ -39,22 +39,44 @@ if (!empty($_POST))
 			$arq_libre=db_execute("UPDATE arqs_testbedims SET activo='F', usuario='libre' WHERE id='" . $id . "'");
 			echo($arq_libre);
 			break;
-		case '3':
-			$sql = "INSERT INTO arqs_testbedims (arquitectura, dominio, activo, usuario, descripcion, imagen) VALUES ('".$_POST['name_arq']."','".$_POST['dominio_arq']."','F','libre', '".$_POST['desc_arq']."','".$_FILES['image']['name']."')";
-			$agregar=db_execute($sql);
-			if ( $agregar == '1') {
-				$id_arq=db_fetch_cell_prepared("SELECT id from arqs_testbedims order by id desc limit 1");
-				echo(return_file_arq($id_arq, $_POST['name_arq'], $_POST['dominio_arq'], $_POST['desc_arq'], $_FILES['image']['name']));
-				//verificar si la arquitectura esta en funcionamiento
-			}
+		case '3'://agregar arquitectura
+			// verificar si existe el archivo
+			if (file_exists('images/images_testbed/images_ims/'.$_FILES['image']['name'])) {
+				if (move_uploaded_file($_FILES['image']['tmp_name'], 'images/images_testbed/images_ims/temp/'.$_FILES['image']['name'])) {
+					$new_name = $now . $_FILES['image']['name'];
+					rename('images/images_testbed/images_ims/temp/'.$_FILES['image']['name'], 'images/images_testbed/images_ims/'.$new_name );
 
-			if (move_uploaded_file($_FILES['image']['tmp_name'], 'images/images_testbed/images_ims/'.$_FILES['image']['name'])) {
-					$up2=db_execute("UPDATE arqs_testbedims SET imagen='".$_FILES['image']['name']."' WHERE id='" . $id_arq . "'");
+					$sql = "INSERT INTO arqs_testbedims (arquitectura, dominio, activo, usuario, descripcion, imagen) VALUES ('".$_POST['name_arq']."','".$_POST['dominio_arq']."','F','libre', '".$_POST['desc_arq']."','".$new_name."')";
+					$agregar=db_execute($sql);
+					if ( $agregar == '1') {
+						$id_arq=db_fetch_cell_prepared("SELECT id from arqs_testbedims order by id desc limit 1");
+						echo(return_file_arq($id_arq, $_POST['name_arq'], $_POST['dominio_arq'], $_POST['desc_arq'], $new_name));
+						//verificar si la arquitectura esta en funcionamiento
+					}
+
 				}else{
 					echo('upload fallo');
 				}
-			break;//agregar arquitectura
-		case '4':
+					
+			}else{
+				if (move_uploaded_file($_FILES['image']['tmp_name'], 'images/images_testbed/images_ims/'.$_FILES['image']['name'])) {
+						$up2=db_execute("UPDATE arqs_testbedims SET imagen='".$_FILES['image']['name']."' WHERE id='" . $id_arq . "'");
+					}else{
+						echo('upload fallo');
+					}
+					$sql = "INSERT INTO arqs_testbedims (arquitectura, dominio, activo, usuario, descripcion, imagen) VALUES ('".$_POST['name_arq']."','".$_POST['dominio_arq']."','F','libre', '".$_POST['desc_arq']."','".$_FILES['image']['name']."')";
+					$agregar=db_execute($sql);
+					if ( $agregar == '1') {
+						$id_arq=db_fetch_cell_prepared("SELECT id from arqs_testbedims order by id desc limit 1");
+						echo(return_file_arq($id_arq, $_POST['name_arq'], $_POST['dominio_arq'], $_POST['desc_arq'], $_FILES['image']['name']));
+						//verificar si la arquitectura esta en funcionamiento
+					}
+			}
+			
+
+			
+			break;
+		case '4'://eliminar arquitectura
 			$file_name=db_fetch_cell_prepared("select imagen from arqs_testbedims where id='".$id."'");
 			$eliminar=db_execute("DELETE FROM arqs_testbedims WHERE id='".$id."'");
 			if ( $eliminar == '1') {				
@@ -65,22 +87,39 @@ if (!empty($_POST))
 		
 		case '5': //modificar arqutiectura
 			$error=0;
-			$up=db_execute("UPDATE arqs_testbedims SET arquitectura = '" . $_POST['name_arq'] . "', dominio ='" . $_POST['dominio_arq'] . "', activo='F', usuario ='libre', descripcion='".$_POST['desc_arq']."', imagen='".$_FILES['image']['name']."' WHERE id='" . $id . "'");	
+
+			$file_name=db_fetch_cell_prepared("select imagen from arqs_testbedims where id='".$id."'");
+			if ($_FILES['image']['name'] == '') {
+				$up=db_execute("UPDATE arqs_testbedims SET arquitectura = '" . $_POST['name_arq'] . "', dominio ='" . $_POST['dominio_arq'] . "', activo='F', usuario ='libre', descripcion='".$_POST['desc_arq']."' WHERE id='" . $id . "'");
+			}else{
+				
+				
+				if (file_exists('images/images_testbed/images_ims/'.$_FILES['image']['name'])) {
+					if (move_uploaded_file($_FILES['image']['tmp_name'], 'images/images_testbed/images_ims/temp/'.$_FILES['image']['name'])) {
+						$new_name = $now . $_FILES['image']['name'];
+						rename('images/images_testbed/images_ims/temp/'.$_FILES['image']['name'], 'images/images_testbed/images_ims/'.$new_name );
+						$up=db_execute("UPDATE arqs_testbedims SET arquitectura = '" . $_POST['name_arq'] . "', dominio ='" . $_POST['dominio_arq'] . "', activo='F', usuario ='libre', descripcion='".$_POST['desc_arq']."', imagen='".$new_name."' WHERE id='" . $id . "'");
+					}
+				
+				}else{
+					$up=db_execute("UPDATE arqs_testbedims SET arquitectura = '" . $_POST['name_arq'] . "', dominio ='" . $_POST['dominio_arq'] . "', activo='F', usuario ='libre', descripcion='".$_POST['desc_arq']."', imagen='".$_FILES['image']['name']."' WHERE id='" . $id . "'");
+				if (move_uploaded_file($_FILES['image']['tmp_name'], 'images/images_testbed/images_ims/'.$_FILES['image']['name'])) {
+						$up2=db_execute("UPDATE arqs_testbedims SET imagen='".$_FILES['image']['name']."' WHERE id='" . $id . "'");
+					}else{
+						echo($error++);
+					}
+				}
+			}
 			if ( $up == '1') {
 
 				echo("<tr id='line".$id."'>
 			 		<td>".$_POST['name_arq']."</td>
 			 		<td>".$_POST['dominio_arq']."</td>
 			 		<td class='edisplay'>".$_POST['desc_arq']."</td>
-			 		<td class='edisplay'>".$_FILES['image']['name']."</td>
+			 		<td class='edisplay'>".$file_name."</td>
 			 		<td>libre</td>
 			 		<td><button class='btn_arq_action' id='btn_liberar".$id."' name='liberar' style='background:green;'> <i class='fa fa-unlock fa-lg'></i></button><button class='btn_arq_action' id='btn_editar".$id."'name='editar' style='background:blue;'> <i class='fa fa-edit fa-lg'></i></button><button class='btn_arq_action' id='btn_eliminar".$id."' name='eliminar' style='background:red;'> <i class='fa fa-trash fa-lg'></i></button></td></tr>");
-				// subir la imagen
-				if (move_uploaded_file($_FILES['image']['tmp_name'], 'images/images_testbed/images_ims/'.$_FILES['image']['name'])) {
-						$up2=db_execute("UPDATE arqs_testbedims SET imagen='".$_FILES['image']['name']."' WHERE id='" . $id . "'");
-					}else{
-						echo($error++);
-					}
+				
 
 			}else{
 				echo($error++);
